#165 Reorganising some code and imports, removing models.py
