--- webcompat/models.py
@@ -1,34 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-from db import session_db
-from db import session_engine
-from hashlib import sha512
-from sqlalchemy import Column
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Integer
-from sqlalchemy import String
-from uuid import uuid4
-
-Base = declarative_base()
-Base.query = session_db.query_property()
-
-
-class User(Base):
-    __tablename__ = 'users'
-
-    user_id = Column(String(128), unique=True, primary_key=True)
-    access_token = Column(String(128), unique=True)
-
-    def __init__(self, access_token):
-        self.access_token = access_token
-        # We use the user_id in the session cookie to identify auth'd users.
-        # Here we salt and hash the GitHub access token so you can't get
-        # back to the auth token if the session cookie was ever compromised.
-        self.user_id = sha512(access_token + uuid4().hex).hexdigest()[0:128]
-
-
-Base.metadata.create_all(bind=session_engine)
